Update Some F-words
## Extended Description
This is and is not fiction (determined by relativity/perspective),  
but is objectively a part of the fiction **Fiction Slayer Punk** "on" Earth “in” **Earth Reality T-626**  
“in” **Multiverse T-626** “in” **multi-multiverse T-626**  
(Your "current" reality if you are reading this).  

This work is part of the public domain under **CC0 1.0 Universal (Public Domain Dedication)**.  
To the extent possible under law, the creator(s) have waived all copyright and related or neighboring rights.  
Feel free to copy, modify, distribute, and perform the work, even for commercial purposes.